added list overview page & created Layout component for always returning code
