+++ pages/edit-entry/[id].js
@@ -1,4 +1,3 @@
-import Head from "next/head";
 import styled from "styled-components";
 import { useState } from "react";
 import { useRouter } from "next/router";
@@ -6,7 +5,7 @@ import { useCategoriesStore, useListItemsStore } from "../../useStore";
 import { nanoid } from "nanoid";
 
 // Components
-import Header from "../../components/Header";
+import Layout from "../../components/Layout";
 import {
   FormMain,
   StyledForm,
@@ -158,13 +157,7 @@ export default function EditEntry() {
 
   return (
     <>
-      <Head>
-        <title>Bearbeite list-item</title>
-        <meta name="description" content="JustList App" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <Header>Bearbeiten</Header>
+      <Layout>Bearbeiten</Layout>
       {currentInfo[1] && <Info>{currentInfo[1]}</Info>}
       <FormMain>
         <StyledForm onSubmit={(event) => handleSubmit(event)}>
